requirements.txt
openai-whisper==20240930
torch>=1.9.0
torchaudio>=0.9.0
numpy<2.0
ffmpeg-python>=0.2.0
PySide6>=6.0.0
pygame>=2.0.0
mutagen>=1.45.0
soundfile
Markdown
Pygments>=0.10.0
requests>=2.25.0
